backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
supabase>=2.7.0
python-dotenv>=1.0.1
pydantic>=2.7.1
python-multipart>=0.0.9
httpx>=0.27.0
opencv-python-headless>=4.9.0
pytesseract>=0.3.10
openai>=1.30.0
Pillow>=10.3.0
numpy>=1.26.0
